Project init

+++ ewmmain-service/src/main/java/ru/practicum/statmain/category/CategoryService.java
@@ -5,6 +5,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import ru.practicum.statmain.category.dto.CategoryCreateRequest;
 import ru.practicum.statmain.category.dto.CategoryDto;
+import ru.practicum.statmain.exception.ConflictException;
+import ru.practicum.statmain.exception.NotFoundException;
 
 import java.util.List;
 
@@ -15,29 +17,43 @@ public class CategoryService {
     private final CategoryRepository categoryRepository;
 
     public CategoryDto addCategory(CategoryCreateRequest request) {
+        checkName(request.getName());
         Category category = CategoryMapper.toEntity(request);
         categoryRepository.save(category);
         return CategoryMapper.toResponse(category);
     }
 
     public void deleteCategory(Integer id) {
+        Category category = categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Category not found"));
+        if (!category.getEvents().isEmpty()) {
+            throw new ConflictException("Category has events");
+        }
         categoryRepository.deleteById(id);
     }
 
     public CategoryDto updateCategory(Integer id, CategoryCreateRequest request) {
         Category category = categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Category not found"));
+        if (!request.getName().equals(category.getName())) {
+            checkName(request.getName());
         category.setName(request.getName());
         categoryRepository.save(category);
+        }
         return CategoryMapper.toResponse(category);
     }
 
     public CategoryDto getCategory(Integer id) {
         return CategoryMapper.toResponse(categoryRepository.findById(id).orElseThrow(
-                () -> new RuntimeException("Category not found")));
+                () -> new NotFoundException("Category not found")));
     }
 
     public List<CategoryDto> getCategories(Integer from, Integer size) {
         return CategoryMapper.toResponse(categoryRepository.findAll(PageRequest.of(from / size, size)).getContent());
+    }
+
+    private void checkName(String name) {
+        if (categoryRepository.findByName(name).isPresent()) {
+            throw new ConflictException("Category with name " + name + " already exists");
+        }
     }
 
 }
